requirements.txt
streamlit
streamlit-folium
tensorflow
pillow
folium
numpy
